feat: Implement cognitive objects architecture (#83)
This PR lays the foundation for the "stay organized effortlessly" part of our "switch tasks immediately, stay organized effortlessly" value proposition. The next step is building tools to populate the cognitive fields, and UI to gracefully compose and recompose information environments with no surprises.

Note that most of the operations are currently taking place on "webpage" objects. It's important to quickly demonstrate that other media types (emails, pdfs, images, videos, charts) can also become objects and slices to illustrate that we are creating an environment with browser as infrastructure, not a browser.